+++ spec/extensions/regexp_spec.rb
@@ -2,6 +2,34 @@ require 'spec_helper'
 require 'ronin/extensions/regexp'
 
 describe Regexp do
+  describe "WORD" do
+    let(:word) { 'dog' }
+
+    subject { Regexp::WORD }
+
+    it "should not match single letters" do
+      subject.match('A').should be_nil
+    end
+
+    it "should not include ending periods" do
+      subject.match("#{word}.")[0].should == word
+    end
+
+    it "should not include leading punctuation" do
+      subject.match("'#{word}")[0].should == word
+    end
+
+    it "should not include tailing punctuation" do
+      subject.match("#{word}'")[0].should == word
+    end
+
+    it "should include punctuation in the middle of the word" do
+      name = "O'Brian"
+
+      subject.match(name)[0].should == name
+    end
+  end
+
   describe "OCTET" do
     subject { Regexp::OCTET }