+++ ChangeLog.md
@@ -8,6 +8,7 @@
 * Added {Ronin::Binary::Hexdump::Parser}.
 * Added {Ronin::Fuzzing::Template}.
 * Added {Ronin::Fuzzing::Repeater}.
+* Added {Ronin::Fuzzing::Mutator}.
 * Added {Ronin::Wordlist.create}.
 * Added {Ronin::Wordlist#path} and {Ronin::Wordlist#words}.
 * Added {Ronin::Wordlist#save}.